extract NPX script into its own package
Extract NPX script wrapper logic from the `compatibility` package to its own `script` package. This will allow re-use of `compatibility` package in the GH action without pulling any unnecessary dependencies.
