+++ .editorconfig
@@ -17,6 +17,9 @@ indent_size = 2
 [*.{yml,yaml}]
 indent_size = 2
 
+[*.json]
+indent_size = 2
+
 [.prettierrc]
 indent_size = 2
