+++ package.json
@@ -1,45 +1,44 @@
 {
-    "name": "@apollo/federation-subgraph-compatibility-monorepo",
-    "description": "Monorepo for Apollo Federation Subgraph Compatibility Testing packages",
-    "bugs": {
-        "url": "https://github.com/apollographql/apollo-federation-subgraph-compatibility/issues"
-    },
-    "homepage": "https://github.com/apollographql/apollo-federation-subgraph-compatibility#readme",
-    "author": "Apollo <packages@apollographql.com>",
-    "license": "MIT",
-    "workspaces": [
-        "packages/*",
-        "packages/subgraphs/*"
-    ],
-    "scripts": {
-        "build": "tsc --build tsconfig.json && npm run package-subgraphs",
-        "compatibility:report": "node packages/compatibility/dist/reportGenerator.js",
-        "compatibility:test": "node packages/compatibility/dist/compatibilityTest.js",
-        "lint": "eslint .",
-        "package-subgraphs": "concurrently \"npm:package-subgraphs:*\"",
-        "package-subgraphs:inventory": "ncc build packages/subgraphs/inventory/index.js --source-map --out packages/compatibility/dist/subgraphs/inventory && cp packages/subgraphs/inventory/Dockerfile packages/subgraphs/inventory/inventory.graphql packages/compatibility/dist/subgraphs/inventory",
-        "package-subgraphs:users": "ncc build packages/subgraphs/users/index.js --source-map --out packages/compatibility/dist/subgraphs/users && cp packages/subgraphs/users/Dockerfile packages/subgraphs/users/users.graphql packages/compatibility/dist/subgraphs/users",
-        "prettier-check": "prettier --check .",
-        "prettier-fix": "prettier --write .",
-        "postinstall": "npm run build"
-    },
-    "main": "dist/index.js",
-    "types": "dist/index.d.ts",
-    "dependencies": {},
-    "devDependencies": {
-        "@types/jest": "^27.5.1",
-        "@typescript-eslint/eslint-plugin": "^5.46.0",
-        "@typescript-eslint/parser": "^5.46.0",
-        "@vercel/ncc": "0.36.0",
-        "concurrently": "^7.2.1",
-        "eslint": "^8.29.0",
-        "eslint-plugin-import": "^2.26.0",
-        "prettier": "^2.8.1",
-        "typescript": "^4.7.2"
-    },
-    "release": {
-        "branches": [
-            "main"
-        ]
-    }
+  "name": "@apollo/federation-subgraph-compatibility-monorepo",
+  "description": "Monorepo for Apollo Federation Subgraph Compatibility Testing packages",
+  "bugs": {
+    "url": "https://github.com/apollographql/apollo-federation-subgraph-compatibility/issues"
+  },
+  "homepage": "https://github.com/apollographql/apollo-federation-subgraph-compatibility#readme",
+  "author": "Apollo <packages@apollographql.com>",
+  "license": "MIT",
+  "workspaces": [
+    "packages/*",
+    "packages/subgraphs/*"
+  ],
+  "scripts": {
+    "build": "tsc --build tsconfig.json && npm run package-subgraphs",
+    "clean": "rm -rf packages/*/dist packages/*/tsconfig.tsbuildinfo",
+    "compatibility:report": "node packages/script/dist/generateReportCommand.js",
+    "compatibility:test": "node packages/script/dist/compatibilityTestCommand.js",
+    "lint": "eslint .",
+    "package-subgraphs": "concurrently \"npm:package-subgraphs:*\"",
+    "package-subgraphs:inventory": "ncc build packages/subgraphs/inventory/index.js --source-map --out packages/compatibility/dist/subgraphs/inventory && cp packages/subgraphs/inventory/Dockerfile packages/subgraphs/inventory/inventory.graphql packages/compatibility/dist/subgraphs/inventory",
+    "package-subgraphs:users": "ncc build packages/subgraphs/users/index.js --source-map --out packages/compatibility/dist/subgraphs/users && cp packages/subgraphs/users/Dockerfile packages/subgraphs/users/users.graphql packages/compatibility/dist/subgraphs/users",
+    "prettier-check": "prettier --check .",
+    "prettier-fix": "prettier --write .",
+    "postinstall": "npm run build"
+  },
+  "dependencies": {},
+  "devDependencies": {
+    "@types/jest": "^27.5.1",
+    "@typescript-eslint/eslint-plugin": "^5.46.0",
+    "@typescript-eslint/parser": "^5.46.0",
+    "@vercel/ncc": "0.36.0",
+    "concurrently": "^7.2.1",
+    "eslint": "^8.29.0",
+    "eslint-plugin-import": "^2.26.0",
+    "prettier": "^2.8.1",
+    "typescript": "^4.7.2"
+  },
+  "release": {
+    "branches": [
+      "main"
+    ]
+  }
 }
