+++ packages/compatibility/package.json
@@ -1,39 +1,39 @@
 {
-  "name": "@apollo/federation-subgraph-compatibility",
-  "version": "0.0.10",
-  "description": "A CLI tool for checking a subgraph's compatibility with a federated gateway",
-  "bin": {
-    "fedtest": "./dist/compatibilityTest.js"
-  },
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "name": "@apollo/federation-subgraph-compatibility-tests",
+  "version": "0.0.11",
+  "description": "Apollo Federation Subgraph Compatibility tests",
   "author": "Apollo <packages@apollographql.com>",
   "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/apollographql/apollo-federation-subgraph-compatibility.git",
+    "directory": "packages/compatibility/"
+  },
+  "keywords": [
+    "apollo",
+    "federation",
+    "compatibility"
+  ],
   "bugs": {
     "url": "https://github.com/apollographql/apollo-federation-subgraph-compatibility/issues"
   },
-  "homepage": "https://github.com/apollographql/apollo-federation-subgraph-compatibility#readme",
+  "homepage": "https://github.com/apollographql/apollo-federation-subgraph-compatibility/blob/main/packages/compatibility/README.md",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "dependencies": {
     "@apollo/rover": "^0.10.0",
-    "commander": "^9.4.1",
     "debug": "^4.3.4",
     "execa": "^5.1.1",
     "graphql": "^16.6.0",
     "jest": "^28.1.0",
-    "js-yaml": "^4.1.0",
-    "lodash": "^4.17.21",
     "make-fetch-happen": "^10.1.5",
     "pm2": "^5.2.2",
     "ts-jest": "^28.0.3"
   },
   "devDependencies": {
     "@types/debug": "^4.1.7",
-    "@types/js-yaml": "^4.0.5",
-    "@types/lodash": "^4.14.182",
     "@types/make-fetch-happen": "^9.0.2",
-    "@types/node": "^17.0.35",
-    "@types/yargs": "^17.0.10",
-    "concurrently": "^7.2.1"
+    "@types/node": "^17.0.35"
   },
   "release": {
     "branches": [
