+++ packages/compatibility/src/index.ts
@@ -1,0 +1,6 @@
+// export functions
+export { compatibilityTest } from './compatibilityTest';
+export { generateReport } from './reportGenerator';
+
+// export types
+export { DockerConfig, Pm2Config } from './startSupergraph';
