+++ packages/script/package.json
@@ -1,0 +1,38 @@
+{
+  "name": "@apollo/federation-subgraph-compatibility",
+  "version": "0.0.11",
+  "description": "A CLI tool for checking a subgraph's compatibility with a federated gateway",
+  "bin": {
+    "fedtest": "./dist/compatibilityTestCommand.js",
+    "fedReport": "./dist/generateReportCommand.js"
+  },
+  "author": "Apollo <packages@apollographql.com>",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/apollographql/apollo-federation-subgraph-compatibility.git",
+    "directory": "packages/script/"
+  },
+  "keywords": [
+    "apollo",
+    "federation",
+    "compatibility"
+  ],
+  "bugs": {
+    "url": "https://github.com/apollographql/apollo-federation-subgraph-compatibility/issues"
+  },
+  "homepage": "https://github.com/apollographql/apollo-federation-subgraph-compatibility/blob/main/packages/script/README.md",
+  "dependencies": {
+    "@apollo/federation-subgraph-compatibility-tests": "^0.0.11",
+    "commander": "^9.4.1",
+    "debug": "^4.3.4"
+  },
+  "devDependencies": {
+    "@types/debug": "^4.1.7"
+  },
+  "release": {
+    "branches": [
+      "main"
+    ]
+  }
+}
